requirements.txt
pytest==3.0.6
requests==2.13.0
vcrpy==1.10.5
xmltodict==0.10.2
